+++ apps/catalog-marketplace/catalog-marketplace/src/app/movie/search/search-validators.form.ts
@@ -1,5 +1,5 @@
 import { AbstractControl } from '@angular/forms';
-import { movieLanguages } from './marketplace-search.form';
+import { movieLanguages } from './search.form';
 
 export function languageValidator(control: AbstractControl): { [key: string]: boolean } | null {
   return !movieLanguages.includes(control.value) ? { languageNotSupported: true } : null;
