+++ apps/catalog-marketplace/catalog-marketplace/src/app/movie/search/search.component.scss
@@ -1,10 +1,15 @@
-.header-img {
+header {
+  padding: 100px 0;
   // TODO: replace image when the real one is available.
-  background: url('/assets/images/catalog-marketplace-banner.png') no-repeat center center fixed;
+  background-image: url('/assets/images/catalog-marketplace-banner.png');
+  background-position: center center;
+  background-attachment: fixed;
   background-size: cover;
-  width: 100%;
-  height: 400px;
-  margin-bottom: 20px;
+
+}
+
+section {
+  padding: 20px 100px;
 }
 
 mat-divider {
@@ -33,5 +38,4 @@ mat-divider {
 movie-display-card-item {
   width: 320px;
   margin-bottom: 10px;
-  cursor: pointer;
 }
